test(01-02): add failing RDR-10c tests for precursor/scan join (RED)
- Add precursor_ms2_index2_has_one_precursor: H1 join by source_index VALUE
- Add precursor_isolation_window_index2: isolation target/offsets ground truth
- Add precursor_activation_parameters_index2: MS:1000133 CID + MS:1000045 CE "35"
- Add precursor_selected_ion_index2: selected_ion_mz 810.789428, intensity 1994039.125
- Add ms1_spectrum_has_empty_precursors: MS1 index 0 has empty precursors
- Add multirow_sweep_ms2_carry_precursor_ms1_carry_none: exactly 34 MS2 / 14 MS1
- Add h2_selected_ion_attach_by_precursor_index: attach by (source_index, precursor_index)
- Add scan_parameters_ms1000800_present: MS:1000800 integer-arm value "100000"
- Add scan_windows_present_on_ms1_spectrum: window 200-2000 on index 0
- Add scan_windows_source_index_join_correct_for_two_spectra: H1 join for scan
- Fix ScanWindow.lower/upper_limit to float (matches float32 Arrow schema)

+++ include/mzpeak/spectrum_metadata.h
@@ -101,13 +101,17 @@ struct PrecursorInfo {
 
 /**
  * One scan window (m/z range) within a scan's scan_windows list.
+ * The Parquet schema stores lower/upper limits as float32; the type matches
+ * the actual Arrow column type to avoid a narrowing conversion.
  */
 struct ScanWindow {
-  /// `MS_1000501_scan_window_lower_limit` — lower m/z bound (double).
-  std::optional<double> lower_limit;
+  /// `MS_1000501_scan_window_lower_limit_unit_MS_1000040` — lower m/z bound
+  /// (float, as stored in the Parquet schema).
+  std::optional<float> lower_limit;
 
-  /// `MS_1000500_scan_window_upper_limit` — upper m/z bound (double).
-  std::optional<double> upper_limit;
+  /// `MS_1000500_scan_window_upper_limit_unit_MS_1000040` — upper m/z bound
+  /// (float, as stored in the Parquet schema).
+  std::optional<float> upper_limit;
 
   /// Additional CV parameters on this scan window.
   std::vector<CvParam> parameters;
